Mastication Arduino Sensor
This change adds the mastication Arduino sensor code to the project.

The code runs all potentiometer values through an averager, a direction
switch detector, and finally a chomp detector. Upon detecting a chomp,
it will output "Detected chomp" and the count of the session.

Tested by running on the masticator mask and validated that it was
detecting chomps correctly.